Refactor operators into separate modules
